requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
transformers[torch]==4.35.0
spacy==3.7.2
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
huggingface-hub==0.17.3
protobuf==4.24.4
numpy==1.26.0
python-dotenv==1.0.0
redis==5.0.1
datasets==2.14.5
scikit-learn==1.3.0
